pyproject.toml
[project]
name = "amarillo"
version = "0.0.15a3"
description = "Aggregates and enhances carpooling-offers and publishes them as GTFS(-RT)"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["amarillo", "ridesharing", "carpooling", "gtfs", "gtfs-rt"]
dependencies = [
    "fastapi[all]==0.109.0",
    "geopandas==0.14",
    "uvicorn[standard]==0.23.2",
    "pydantic[dotenv]==2.4.2",
    "protobuf==3.20.3",
    "starlette~=0.35",
    "requests==2.31.0",
    "pyproj==3.6.1",
    "geojson-pydantic==1.0.1",
    "watchdog==3.0.0",
    "python-jose[cryptography]",
    "bcrypt==4.0.1",
    "passlib[bcrypt]"
]

[tool.setuptools.packages]
find = {}

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120
skip-string-normalization = 1

[tool.ruff]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "Q", # pydocstyle
    "T", # mypy
    "R", # pylint
    "S", # flake8
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]

ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "F401", # imported but unused
    "S101", # use of assert detected
    "S608", # TODO: Appropriate SQL injection counter measures should be discussed
]

fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 120

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "single"
docstring-quotes = "double"

[[tool.mypy.overrides]]
# See https://github.com/HBNetwork/python-decouple/issues/122
module = ["dagster_docker", "geopandas", "geopandas.array", "osgeo", "osgeo_utils.samples", "defusedxml", "defusedxml.ElementTree"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# See https://github.com/HBNetwork/python-decouple/issues/122
module = ["geopandas"]

